prevent unnecessary redundant preloading of function/web apps
